Serious bug in cdogs_abstract_template_top:
Have maybe fixed it - problem was with \abaffil0 macro. But now need to fix script that calls it to write the affiliation to the .tex file. I think all I need to do is remove the $$ from the affiliation line.

Also, need to update the fancyheadings package to fancyhdr to follow the warnings.